% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_threads}
\alias{set_threads}
\title{Set number of threads used by holtz}
\usage{
set_threads(n)
}
\arguments{
\item{n}{Number of threads to be used by RcppParallel}
}
\description{
Set number of threads used by holtz
}
